moving isValidRawRule to its own file so we can remove it later on
